refactor: move define to builtins(close: #867)
