+++ crates/rspack_core/src/options/compiler_options.rs
@@ -1,5 +1,5 @@
 use crate::{
-  Builtins, BundleEntries, Context, Define, DevServerOptions, Devtool, External, ExternalType,
+  Builtins, BundleEntries, Context, DevServerOptions, Devtool, External, ExternalType,
   ModuleOptions, OutputOptions, Plugins, Resolve, Target,
 };
 
@@ -14,7 +14,6 @@ pub struct CompilerOptions {
   pub builtins: Builtins,
   pub plugins: Plugins,
   pub module: ModuleOptions,
-  pub define: Define,
   pub devtool: Devtool,
   pub external: Vec<External>,
   pub external_type: ExternalType,
