+++ crates/rspack_loader_sass/tests/fixtures.rs
@@ -51,7 +51,6 @@ async fn loader_test(actual: impl AsRef<Path>, expected: impl AsRef<Path>) {
           builtins: Default::default(),
           plugins: Default::default(),
           module: Default::default(),
-          define: Default::default(),
           external: Default::default(),
           external_type: ExternalType::Auto,
           emit_error: false,
