+++ crates/rspack_plugin_javascript/tests/fixtures/builtins/define/expected/runtime.js
