+++ crates/rspack_plugin_javascript/tests/fixtures/builtins/define/lib.js
